Switch to mocha for unit tests (#2525)
* Switch to motcha for unit tests
