+++ package.json
@@ -7,7 +7,7 @@
     "test": "npm run check-types && npm run lint && npm run check-formatting && npm run test-log-conventions && npm run test-unit && npm run test-misc",
     "test-misc": "node ./bin/test_package_lock_version",
     "test-log-conventions": "node ./bin/log-conventions",
-    "test-unit": "node test/unit",
+    "test-unit": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha -r ts-node/register 'test/mocha/**/*.ts'",
     "test-e2e-ci": "env NODE_ENV=test testcafe --skip-js-errors \"electron:./\" \"./test/testcafe/test.js\"",
     "test-e2e": "env NODE_ENV=test testcafe --skip-js-errors \"electron:./\" \"./test/testcafe/test.js\"",
     "check-types": "tsc -b src/check-all-tsconfig.json",
@@ -111,10 +111,12 @@
     "@babel/core": "^7.7.7",
     "@babel/preset-env": "^7.7.7",
     "@babel/preset-react": "^7.7.4",
+    "@types/chai": "^4.3.0",
     "@types/debounce": "^1.2.0",
     "@types/emoji-mart": "^3.0.2",
     "@types/mapbox-gl": "^0.54.5",
     "@types/mime-types": "^2.1.0",
+    "@types/mocha": "^9.0.0",
     "@types/node": "^14.14.20",
     "@types/node-fetch": "^2.5.7",
     "@types/prop-types": "^15.7.3",
@@ -126,6 +128,7 @@
     "@types/url-parse": "^1.4.3",
     "@typescript-eslint/eslint-plugin": "^5.3.1",
     "@typescript-eslint/parser": "^5.3.1",
+    "chai": "^4.3.4",
     "electron": "^13.1.6",
     "electron-builder": "^22.12.0",
     "electron-devtools-installer": "^3.1.1",
@@ -136,12 +139,13 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "glob-watcher": "^5.0.5",
     "hallmark": "^4.0.0",
+    "mocha": "^9.1.4",
     "prettier": "2.2.1",
     "sass": "^1.26.5",
-    "tape": "^4.12.0",
     "testcafe": "^1.15.0",
     "testcafe-browser-provider-electron": "0.0.16",
     "testcafe-react-selectors": "^4.1.5",
+    "ts-node": "^10.4.0",
     "typescript": "^4.4.4",
     "walk": "^2.3.14",
     "xml-js": "^1.6.8"
