+++ src/shared/logger.ts
@@ -125,6 +125,7 @@ function getStackTrace(): StackFrame[] | string {
 }
 
 export class Logger {
+  //@ts-ignore
   isMainProcess = typeof window === 'undefined'
   constructor(public readonly channel: string) {
     if (channel === 'core/event') {
